HomeScreen - delete expenses
- Added files :HomeScreen.kt, HomeUiData.kt,
- moved colors to colors.kt
- cleaned up some of mainactivity( only data and variables are in there)
