+++ app/src/main/java/com/example/budgetapp/ui/theme/Color.kt
@@ -9,3 +9,7 @@ val Pink80 = Color(0xFFEFB8C8)
 val Purple40 = Color(0xFF6650a4)
 val PurpleGrey40 = Color(0xFF625b71)
 val Pink40 = Color(0xFF7D5260)
+
+val lightPurpleColor = Color(97, 111, 219)
+val lightTealColor = Color(207, 220, 225)
+val TealColor = Color(115, 170, 197)
