+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,6 @@
 <resources>
     <string name="app_name">Budget App</string>
+    <string name="expense">Expense</string>
+    <string name="category">Category</string>
+    <string name="amount">$ </string>
 </resources>
